#86et9ae91 - feat: refactor live session form components and introduce modal for session creation; enhance user experience with improved structure and validation logic
